apps/api/pyproject.toml
[project]
name = "docsai-api"
version = "0.1.0"
description = "FastAPI backend for DocsAI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "asyncpg>=0.29.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.27.0",
  "structlog>=24.4.0",
  "celery[redis]>=5.4.0",
  "redis>=5.0.0",
  "supabase>=2.6.0",
  "openai>=1.40.0",
  "pymupdf>=1.24.0"
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
